+++ packages/cryptoplease/lib/app/screens/authenticated/flow.dart
@@ -4,8 +4,8 @@ import 'package:cryptoplease/core/accounts/bl/accounts_bloc.dart';
 import 'package:cryptoplease/core/balances/bl/balances_bloc.dart';
 import 'package:cryptoplease/core/conversion_rates/bl/conversion_rates_bloc.dart';
 import 'package:cryptoplease/core/conversion_rates/module.dart';
+import 'package:cryptoplease/core/split_key_payments/transaction/tx_creator_selector.dart';
 import 'package:cryptoplease/core/user_preferences.dart';
-import 'package:cryptoplease/data/transaction/tx_creator_selector.dart';
 import 'package:cryptoplease/features/add_funds/module.dart';
 import 'package:cryptoplease/features/airdrop/module.dart';
 import 'package:cryptoplease/features/backup_phrase/module.dart';
