+++ packages/cryptoplease/lib/core/split_key_payments/transaction/creators/solana_tx_creator.dart
@@ -2,8 +2,8 @@ import 'package:collection/collection.dart';
 import 'package:cryptoplease/config.dart';
 import 'package:cryptoplease/core/accounts/bl/account.dart';
 import 'package:cryptoplease/core/solana_helpers.dart';
+import 'package:cryptoplease/core/split_key_payments/transaction/tx_creator.dart';
 import 'package:cryptoplease/core/tokens/token.dart';
-import 'package:cryptoplease/data/transaction/tx_creator.dart';
 import 'package:cryptoplease/features/incoming_split_key_payment/bl/tx_processor.dart';
 import 'package:cryptoplease/features/outgoing_transfer/bl/outgoing_payment.dart';
 import 'package:dfunc/dfunc.dart';
