+++ packages/cryptoplease/lib/main.dart
@@ -4,11 +4,11 @@ import 'package:cryptoplease/config.dart';
 import 'package:cryptoplease/core/accounts/module.dart';
 import 'package:cryptoplease/core/analytics/analytics_manager.dart';
 import 'package:cryptoplease/core/balances/module.dart';
+import 'package:cryptoplease/core/split_key_payments/transaction/creators/cp_tx_creator.dart';
+import 'package:cryptoplease/core/split_key_payments/transaction/creators/solana_tx_creator.dart';
+import 'package:cryptoplease/core/split_key_payments/transaction/tx_creator_selector.dart';
 import 'package:cryptoplease/core/tokens/token_list.dart';
 import 'package:cryptoplease/data/db/db.dart';
-import 'package:cryptoplease/data/transaction/creators/cp_tx_creator.dart';
-import 'package:cryptoplease/data/transaction/creators/solana_tx_creator.dart';
-import 'package:cryptoplease/data/transaction/tx_creator_selector.dart';
 import 'package:cryptoplease/features/incoming_split_key_payment/module.dart';
 import 'package:cryptoplease/features/outgoing_transfer/module.dart';
 import 'package:cryptoplease/features/pending_request/module.dart';
