+++ packages/cryptoplease/test/features/incoming_split_key_payment/bloc_test.dart
@@ -1,9 +1,9 @@
-import 'package:cryptoplease/core/api_version.dart';
+import 'package:cryptoplease/core/split_key_payments/split_key_api_version.dart';
+import 'package:cryptoplease/core/split_key_payments/transaction/creators/cp_tx_creator.dart';
+import 'package:cryptoplease/core/split_key_payments/transaction/creators/solana_tx_creator.dart';
+import 'package:cryptoplease/core/split_key_payments/transaction/tx_creator_selector.dart';
 import 'package:cryptoplease/core/tokens/token.dart';
 import 'package:cryptoplease/core/wallet.dart';
-import 'package:cryptoplease/data/transaction/creators/cp_tx_creator.dart';
-import 'package:cryptoplease/data/transaction/creators/solana_tx_creator.dart';
-import 'package:cryptoplease/data/transaction/tx_creator_selector.dart';
 import 'package:cryptoplease/features/incoming_split_key_payment/bl/bloc.dart';
 import 'package:cryptoplease/features/incoming_split_key_payment/bl/models.dart';
 import 'package:cryptoplease/features/incoming_split_key_payment/bl/tx_processor.dart';
